Crate rust_kb

Rust KB is a library implementing basic knowledge base function in Rust. Rust KB provides an easy interface to create knowledge bases from full knowledge base files, or rule/fact strings individually.

Usage

This crate is not yet on crates.io. Therefore, to use this library, download the source code and add it as a module.

Add this to your crate root:

extern crate rust_kb;

Structs

Fact
Defines a fact relationship between two or more arguments

KnowledgeBase
A data structure which can take in facts and rules, and respond to logical questions and queries

Rule
Defines a knowledge base fact that can be inferred, given 1 or more facts as the premise

Enums

Statement
Abstraction that encompasses facts and rules