Struct rust_kb::KnowledgeBase

pub struct KnowledgeBase { /* fields omitted */ }

A data structure which can take in facts and rules, and respond to logical questions and queries

A knowledge base can take (as well as remove) facts and rules to generate facts which it knows to be true. Then, the knowledge base can be asked logical questions and will answer whether it can determine the answer.

Methods

impl KnowledgeBase

pub fn new() -> KnowledgeBase

Creates a new, empty knowledge base

This function should be used when one is programmatically building up a knowledge base and no knowledge is known before the start of the program. If any static facts or rules are known beforehand, a knowledge base file should be used, and KnowledgeBase::from_file() should create the knowledge base.

Example

use rust_kb::KnowledgeBase;

let mut kb = KnowledgeBase::new();
// kb knows nothing and all asks will return false

pub fn from_file(filename: &str) -> Result<KnowledgeBase, String>

Attempts to create a knowledge base from a given input file

The input file should formatted as follows. The entire file should be surrounded by kb { ... }

Proper knowledge base delimiters

kb {
    ...
}

All facts should follow. Each fact should be on its own line and be prefixed by "fact:". Then, the fact should be left left parenthesis, the predicate, one or more arguments, and finally a right parenthesis

Proper fact syntax

kb {
    fact: (isa cube box)
    fact: (isa box container)

    ...
}

Finally, all rules should follow. Each rule should be on its own line and be prefixed by "rule:". Then the rule should have a list of one or more facts, a right arrow (->), and finally a fact that can be inferred. These facts should use variables to connect arguments from different facts.

Proper knowledge base file format

kb {
    fact: (isa cube box)
    fact: (isa box container)

    rule: ((inst ?x ?y) (isa ?y ?z)) -> (inst ?x ?z)
}

Example

use rust_kb::KnowledgeBase;

let mut kb = KnowledgeBase::from_file("test/test.kb");

// The knowledge base now has all facts and rules from the file

pub fn create_fact(&mut self, fact: &str) -> Result<Statement, String>

Attempts to create a fact from a given string slice.

If the fact is ill-formatted, the function will return an error. In this context, the fact must be terminated by a semicolon.

Example

use rust_kb::KnowledgeBase;

let mut kb = KnowledgeBase::new();
match kb.create_fact("isa(square, rectangle.") {
    Ok(fact) => { /* Will execute this branch */ },
    Err(_) => { /* Will not execute this branch, because of proper format */ },
}

pub fn create_rule(&mut self, rule: &str) -> Result<Statement, String>

Attempts to create a rule from a given string slice.

If the rule is ill-formatted, the function will return an error. In this context, the fact must be terminated by a semicolon.

Example

use rust_kb::KnowledgeBase;

let mut kb = KnowledgeBase::new();
match kb.create_rule("inst(?x, ?z) :- inst(?x, ?y), isa(?y, ?z).") {
    Ok(fact) => { /* Will execute this branch */ },
    Err(_) => { /* Will not execute this branch, because of proper format */ },
}

pub fn assert(&mut self, statement: Statement) -> Result<Statement, String>

Add a fact or rule to the knowledge base

This function will use inference by forward chaining to add implied facts from given rules. An error will be returned if the statement is already present in the knowledge base. Use this error to detect logical errors, or duplicate assertions in code.

Example

use rust_kb::KnowledgeBase;

let mut kb = KnowledgeBase::new();
match kb.create_fact("isa(square, rectangle.") {
    Ok(fact) => { kb.assert(fact); },
    Err(_) => {},
}

pub fn retract(&mut self, statement: Statement) -> Result<(), String>

Remove a fact or rule from the knowledge base.

This function will remove a specific statement from the knowledge base. In addition, it will recursively chain logic to remove other statements that were dependent on the given statement

Statements that still have support from other Fact/Rule pairs will error on retract

Example

use rust_kb::{KnowledgeBase,Fact};

let mut kb = KnowledgeBase::new();

// Add to the knowledge base

if let Ok(fact) = kb.create_fact("isa(square, rectangle.") {
    kb.retract(fact);
}

pub fn ask(&self, fact: &Statement) -> Result<bool, String>

Ask if a specific fact can be proven by the knowledge base

Example

use rust_kb::KnowledgeBase;

let mut kb = KnowledgeBase::new();
if let Ok(fact) = kb.create_fact("isa(square, rectangle.") {
    kb.ask(&fact);
}

pub fn query(&self, s: &Statement) -> Result<Vec<Vec<(Symbol, Symbol)>>, String>

Query a knowledge base to find all possible bindings to variables in the fact

The given fact should contain at least one variable.

Example

use rust_kb::KnowledgeBase;

let mut kb = KnowledgeBase::new();

// Fill the knowledge base

if let Ok(fact) = kb.create_fact("isa(?object, rectangle.") {
    let result = kb.query(&fact);
    // result is now a vector of all bindings found in the knowledge base.
}